feat(rust): share hosted lifecycle wire contract
