+++ crates/openengine-cluster-protocol/src/lib.rs
@@ -10,6 +10,7 @@ pub mod graph;
 mod graph_profile;
 pub mod lifecycle;
 pub mod logs;
+pub mod native_v2_hosted;
 pub mod native_v2_observation;
 pub mod native_v2_run;
 pub mod native_v2_target;
@@ -30,6 +31,7 @@ pub use fault::*;
 pub use graph::*;
 pub use lifecycle::*;
 pub use logs::*;
+pub use native_v2_hosted::*;
 pub use native_v2_observation::*;
 pub use native_v2_run::*;
 pub use native_v2_target::*;
